removed dead code and one ugly hack
