+++ command/contents/comparisons.py
@@ -4,95 +4,12 @@ import logging
 from command.fast_path import FastPosixPath
 from config import HoardConfig
 from contents.hoard import HoardContents, HoardFSObjects
-from contents.hoard_props import HoardFileStatus
-from contents.repo import RepoContents, RepoFSObjects
+from contents.repo import RepoContents
 from contents.repo_props import FileDesc
 from lmdb_storage.file_object import FileObject
 from lmdb_storage.object_store import ObjectStorage
-from lmdb_storage.tree_iteration import zip_trees_dfs, zip_dfs
+from lmdb_storage.tree_iteration import zip_trees_dfs
 from lmdb_storage.tree_structure import TreeObject, add_object
-from util import safe_hex
-
-
-async def sync_fsobject_to_object_storage(
-        env: ObjectStorage, fsobjects: HoardFSObjects, repo_objects: RepoFSObjects, hoard_config: HoardConfig):
-    if env.roots(write=True)["HOARD"].desired is None:
-        # fixme ugly hack to ensure unit tests do not change
-        with env.objects(write=True) as objects:
-            empty_dir = objects.mktree_from_tuples([])
-        env.roots(write=True)["HOARD"].desired = empty_dir
-
-    return
-    raise_exception = False
-
-    old_root_id = env.roots(False)["HOARD"].desired
-
-    all_nondeleted = [
-        (path.as_posix(), FileObject.create(hfo.fasthash, hfo.size))
-        async for path, hfo in fsobjects.in_folder_non_deleted(FastPosixPath("/"))]
-
-    with env.objects(write=True) as objects:
-        current_root_id = objects.mktree_from_tuples(all_nondeleted)
-
-    with env.objects(write=True) as objects:
-        empty_dir = objects.mktree_from_tuples([])
-    if raise_exception and old_root_id != current_root_id:
-        if old_root_id is not None and current_root_id != empty_dir:
-            raise ValueError(f"Hoard root changed: {old_root_id} != {current_root_id}")
-
-    for remote in hoard_config.remotes.all():
-        with env.objects(write=True) as objects:
-            # remote_current_id = objects.mktree_from_tuples([
-            #     (hoard_config.remotes[remote.uuid].mounted_at.joinpath(path).as_posix(),
-            #      FileObject.create(hfo.fasthash, hfo.size))
-            #     for path, hfo in repo_objects.existing()])
-            remote_current_id = objects.mktree_from_tuples([
-                (path.as_posix(), FileObject.create(hfo.fasthash, hfo.size))
-                async for path, hfo in fsobjects.in_folder(FastPosixPath("/"))
-                if hfo.get_status(remote.uuid) in (HoardFileStatus.AVAILABLE, HoardFileStatus.CLEANUP)])
-
-            remote_desired_id = objects.mktree_from_tuples([
-                (path.as_posix(), FileObject.create(hfo.fasthash, hfo.size))
-                # fixme make path absolute
-                async for path, hfo in fsobjects.in_folder(FastPosixPath("/"))
-                if hfo.get_status(remote.uuid) in (
-                    HoardFileStatus.AVAILABLE, HoardFileStatus.GET, HoardFileStatus.COPY, HoardFileStatus.MOVE)])
-
-        root = env.roots(True)[remote.uuid]
-        if remote_current_id != root.current:
-            if root.current is not None and remote_current_id != empty_dir:
-                if raise_exception:
-                    raise ValueError(
-                        f"current tree wrong - {remote.name}: old={safe_hex(root.current)} new={safe_hex(remote_current_id)}\n{dump_tree_diffs(env, root, remote_current_id)}")
-                logging.error(
-                    f"Current tree is not correct: {remote.name}: {safe_hex(remote_current_id)} is not current, current={safe_hex(root.current)}")
-            # assert root.current is None, f"{remote.name}: {safe_hex(remote_current_id)} is not current, current={safe_hex(root.current)}"
-        if remote_desired_id != root.desired:
-            if raise_exception and root.desired is not None and remote_desired_id != empty_dir:
-                dump_tree_diffs(env, root, remote_current_id)
-                raise ValueError(
-                    f"desired tree wrong - {remote.name}: old={safe_hex(root.desired)} new={safe_hex(remote_desired_id)}\n{dump_tree_diffs(env, root, remote_current_id)}")
-            # assert root.desired is None, f"{remote.name}: {safe_hex(remote_desired_id)} is not desired, desired={safe_hex(root.desired)}"
-
-        # fixme fully delete root.current = remote_current_id
-        # root.desired = remote_desired_id
-
-    env.roots(write=True)["HOARD"].desired = current_root_id
-    print(
-        f"Old HEAD: {binascii.hexlify(old_root_id) if old_root_id is not None else 'None'}"
-        f" vs new head {binascii.hexlify(current_root_id)}.")
-    return current_root_id
-
-
-def dump_tree_diffs(env, root, remote_current_id) -> str:
-    with (env.objects(write=False) as objects):
-        diffs = [
-            f"{path}: {diff_type.value}"
-            for path, diff_type, left_id, right_id, should_skip
-            in zip_dfs(objects, "root", root.current, remote_current_id)]
-        return "\n############ dumping diffs of trees ############\n" + \
-            "\n".join(diffs) + \
-            "\n############# end dumping tree ############\n\n"
 
 
 def sync_object_storage_to_recreate_fsobject_and_fspresence(
